requirements.txt
Appium-Python-Client==1.0.2
attrs==20.3.0
iniconfig==1.1.1
packaging==20.9
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.2
pytest-yaml==1.2.1
PyYAML==5.4.1
selenium==3.141.0
toml==0.10.2
urllib3==1.26.3
